+++ services/galexie/internal/config.go
@@ -180,7 +180,6 @@ func (config *Config) GenerateCaptiveCoreConfig(coreBinFromPath string) (ledgerb
 	params := ledgerbackend.CaptiveCoreTomlParams{
 		NetworkPassphrase:  config.StellarCoreConfig.NetworkPassphrase,
 		HistoryArchiveURLs: config.StellarCoreConfig.HistoryArchiveUrls,
-		UseDB:              true,
 	}
 
 	captiveCoreToml, err := ledgerbackend.NewCaptiveCoreTomlFromData(config.SerializedCaptiveCoreToml, params)
@@ -200,7 +199,6 @@ func (config *Config) GenerateCaptiveCoreConfig(coreBinFromPath string) (ledgerb
 		Log:                 logger.WithField("subservice", "stellar-core"),
 		Toml:                captiveCoreToml,
 		UserAgent:           config.UserAgent,
-		UseDB:               true,
 		StoragePath:         config.StellarCoreConfig.StoragePath,
 	}, nil
 }