+++ services/horizon/internal/ingest/main.go
@@ -121,7 +121,6 @@ type Config struct {
 	CaptiveCoreBinaryPath  string
 	CaptiveCoreStoragePath string
 	CaptiveCoreToml        *ledgerbackend.CaptiveCoreToml
-	CaptiveCoreConfigUseDB bool
 	NetworkPassphrase      string
 
 	HistorySession        db.SessionInterface
@@ -317,12 +316,10 @@ func NewSystem(config Config) (System, error) {
 			ledgerbackend.CaptiveCoreConfig{
 				BinaryPath:            config.CaptiveCoreBinaryPath,
 				StoragePath:           config.CaptiveCoreStoragePath,
-				UseDB:                 config.CaptiveCoreConfigUseDB,
 				Toml:                  config.CaptiveCoreToml,
 				NetworkPassphrase:     config.NetworkPassphrase,
 				HistoryArchiveURLs:    config.HistoryArchiveURLs,
 				CheckpointFrequency:   config.CheckpointFrequency,
-				LedgerHashStore:       ledgerbackend.NewHorizonDBLedgerHashStore(config.HistorySession),
 				Log:                   logger,
 				Context:               ctx,
 				UserAgent:             fmt.Sprintf("captivecore horizon/%s golang/%s", apkg.Version(), runtime.Version()),