+++ services/horizon/internal/test/integration/integration.go
@@ -568,7 +568,6 @@ func (i *Test) getDefaultIngestArgs() map[string]string {
 		"stellar-core-binary-path":  i.coreConfig.binaryPath,
 		"captive-core-config-path":  i.coreConfig.configPath,
 		"captive-core-http-port":    "21626",
-		"captive-core-use-db":       "true",
 		"captive-core-storage-path": i.coreConfig.storagePath,
 		"ingest":                    "true"})
 }
@@ -660,14 +659,12 @@ func (i *Test) CreateCaptiveCoreConfig() (ledgerbackend.CaptiveCoreConfig, error
 		HistoryArchiveURLs:  []string{HistoryArchiveUrl},
 		NetworkPassphrase:   i.config.NetworkPassphrase,
 		CheckpointFrequency: CheckpointFrequency, // This is required for accelerated archive creation for integration test
-		UseDB:               true,
 		StoragePath:         i.CurrentTest().TempDir(),
 	}
 
 	tomlParams := ledgerbackend.CaptiveCoreTomlParams{
 		NetworkPassphrase:  i.config.NetworkPassphrase,
 		HistoryArchiveURLs: []string{HistoryArchiveUrl},
-		UseDB:              true,
 	}
 
 	toml, err := ledgerbackend.NewCaptiveCoreTomlFromFile(i.coreConfig.configPath, tomlParams)